+++ functions/src/model/UsersImported.ts
@@ -1,0 +1,9 @@
+export class UsersImported {
+  ownerIdSet: Set<String>;
+  userIdSet: Set<String>;
+
+  constructor(ownerIdSet: Set<String>, userIdSet: Set<String>) {
+    this.ownerIdSet = ownerIdSet;
+    this.userIdSet = userIdSet;
+  }
+}
