+++ functions/src/utils/UserUtils.ts
@@ -1,6 +1,6 @@
-import Wallet from "./model/Wallet";
 import * as admin from "firebase-admin";
 import * as functions from "firebase-functions";
+import Wallet from "../model/Wallet";
 
 const db = admin.firestore();
 const logger = functions.logger;
@@ -32,12 +32,28 @@ export async function getWalletForUser(userId: String): Promise<Wallet | null> {
   }
 }
 
-// Chunk an array into a given size (so for [1, 2, 3, 4] chunked by 2 it will give us [[1, 2], [3, 4]])
-export function sliceIntoChunks(arr: Array<any>, chunkSize: number) {
-  const res = [];
-  for (let i = 0; i < arr.length; i += chunkSize) {
-    const chunk = arr.slice(i, i + chunkSize);
-    res.push(chunk);
+/**
+ * Get the wallet's for all the given users
+ * @param userIdSet The list of id for which we want the wallet's
+ */
+export async function getWalletForUserSet(
+  userIdSet: Set<String>
+): Promise<Set<Wallet>> {
+  try {
+    logger.debug(`Try to get the wallet's for the user ${userIdSet}`);
+    // Access our wallet collection
+    const collection = db.collection("wallet");
+    const foundWallets = new Set<Wallet>();
+    // Execute the query
+    const snapshot = await collection
+      .where("id", "in", userIdSet)
+      .limit(1)
+      .get();
+    // Find document and map them
+    snapshot.forEach((doc) => foundWallets.add(doc.data() as Wallet));
+    return foundWallets;
+  } catch (exception) {
+    logger.warn("Error when searching for the user wallet", exception);
+    return new Set();
   }
-  return res;
 }
